Created new function levelUpAllPrereqSkills()
It does exactly what you should expect. Modified the plus button to call this function if you attempt to level up a locked skill.

TODO: Remove the brightness(20%) filter. It is no longer needed to show the user that they can't allocate points to that particular skill. Since now, it is possible.